third_party/blink/web_tests/fast/dom/MutationObserver/delivery-order-expected.txt

Test that MutationObservers are delivered to in order of creation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS order.length is 10
PASS order[0] is 0
PASS order[1] is 1
PASS order[2] is 2
PASS order[3] is 3
PASS order[4] is 4
PASS order[5] is 5
PASS order[6] is 6
PASS order[7] is 7
PASS order[8] is 8
PASS order[9] is 9
PASS successfullyParsed is true

TEST COMPLETE
